Vexel 0.2.0: templates, new API, faster loading

+++ src/builder.py
@@ -11,21 +11,23 @@ from core_utils import *
 from build_utils import *
 
 def MakeFileSystem():
-	VexelLog("Cleaning...", "INFO")
-	rmdir(output['dir'])
+	if check_dir(output['dir']):
+		VexelLog("Cleaning...", "INFO")
+		rmdir(output['dir'])
 	
 	mkdir(output['res'])
 	mkdir(output['dex'])
 	mkdir(output['classes'])
 	mkdir(output['gen'])
 	mkdir(output['key'])
+	mkdir(output['apk'])
 
 
-def build():
+def build(dir=None):
 	start = time.time()
 	
-	MakeFileSystem()
 	config = load_config(project['vexel_build'])
+	MakeFileSystem()
 	update_manifest(project['manifest'], output['manifest'], config)
 	VexelLog("Compile resources...", "BUILD")
 	run_aapt2_compile()
@@ -43,27 +45,30 @@ def build():
 	convert_dex()
 
 	VexelLog("Add classes.dex...", "BUILD")
-	shutil.copy(path(output['dir'], "temp.apk"), path(output['dir'], "temp_unaligned.apk"))
+	shutil.copy(path(output['apk'], "base.apk"), path(output['apk'], "unaligned.apk"))
 	
 	# Add Dex
 	for dex in find_files(output['dex'], ".dex"):
-		add_to_apk(path(output['dir'], "temp_unaligned.apk"), dex, os.path.basename(dex), False, 0)
+		add_to_apk(path(output['apk'], "unaligned.apk"), dex, os.path.basename(dex), False, 0)
 	
 	# Add User Dynamic Libraries
-	add_to_apk(path(output['dir'], "temp_unaligned.apk"), project['dyn_lib'], "lib")
+	add_to_apk(path(output['apk'], "unaligned.apk"), project['dyn_lib'], "lib")
 	
 	# Add Compiled Dynamic Libraries
-	add_to_apk(path(output['dir'], "temp_unaligned.apk"), output['dyn_lib'], "lib")
+	add_to_apk(path(output['apk'], "unaligned.apk"), output['dyn_lib'], "lib")
 	
 	# Add Java Libraries
-	add_to_apk(path(output['dir'], "temp_unaligned.apk"), project['assets'], "assets")
+	add_to_apk(path(output['apk'], "unaligned.apk"), project['assets'], "assets")
 	
 	VexelLog("Align APK...", "BUILD")
 	align_apk()
 
 	VexelLog("Sign APK...", "BUILD")
-	sign_apk()
+	sign_apk(config)
+	
+	mkdir(output['final'])
+	shutil.copy(path(output['apk'], "final.apk"), path(output['final'], "Debug.apk"))
 	
 	elapsed = round(time.time() - start, 2)
 	VexelLog(f"Build Successful in {elapsed} sec", "INFO")
-	VexelLog(f"Apk saved at: {path(output['dir'], 'Debug.apk')}", "INFO")
+	VexelLog(f"Apk saved at: {path(output['final'], 'Debug.apk')}", "INFO")
